SOCINT-117 Code review changes and common update
