+++ pom.xml
@@ -23,7 +23,7 @@
     <groupId>ons.sdc.int.common</groupId>
     <artifactId>common</artifactId>
     <!-- change this to the version of ALL common libs to be used by this build -->
-    <version>1.0.34-SNAPSHOT</version>
+    <version>1.0.35</version>
   </parent>
 
   <scm>
