+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/config/QueueConfig.java
@@ -6,12 +6,8 @@ import lombok.Data;
 @Data
 public class QueueConfig {
   private String caseSubscription;
-  private String caseTopic;
   private String uacSubscription;
-  private String uacTopic;
   private String surveySubscription;
-  private String surveyTopic;
   private String collectionExerciseSubscription;
-  private String ollectionExerciseTopic;
   private Set<String> qidFilterPrefixes;
 }
