+++ src/main/resources/application.yml
@@ -70,13 +70,9 @@ spring:
 
 queueconfig:
   case-subscription: event_case-update_rh
-  case-topic: event_case-update
   uac-subscription: event_uac-update_rh
-  uac-topic: event_uac-update
   collection-exercise-subscription: event_collection-exercise-update_rh
-  collection-exercise-topic: event_collection-exercise-update
   survey-subscription: event_survey-update_rh
-  survey-topic: event_survey-update
   qid-filter-prefixes:
 
 messaging:
