+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/message/impl/CollectionExerciseEventReceiverImplIT_Test.java
@@ -35,8 +35,7 @@ import uk.gov.ons.ctp.integration.rhsvc.repository.impl.RespondentDataRepository
 /** Spring Integration test of flow received from Response Management */
 @SpringBootTest
 @EnableConfigurationProperties
-@ContextConfiguration(
-    classes = {AppConfig.class, CollectionExerciseEventReceiverImplIT_Config.class})
+@ContextConfiguration(classes = {AppConfig.class, MessageIT_Config.class})
 @ExtendWith(SpringExtension.class)
 @ActiveProfiles("mocked-connection-factory")
 @ResourceLock(value = ParallelTestLocks.SPRING_TEST, mode = READ_WRITE)
