+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/message/impl/MessageIT_Config.java
@@ -8,12 +8,33 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.test.util.ReflectionTestUtils;
 import uk.gov.ons.ctp.integration.rhsvc.config.AppConfig;
 import uk.gov.ons.ctp.integration.rhsvc.config.InboundEventIntegrationConfig;
+import uk.gov.ons.ctp.integration.rhsvc.event.impl.CaseEventReceiverImpl;
+import uk.gov.ons.ctp.integration.rhsvc.event.impl.CollectionExerciseEventReceiverImpl;
+import uk.gov.ons.ctp.integration.rhsvc.event.impl.SurveyEventReceiverImpl;
 import uk.gov.ons.ctp.integration.rhsvc.event.impl.UACEventReceiverImpl;
 
 @Profile("mocked-connection-factory")
 @Configuration
 @Import({InboundEventIntegrationConfig.class, EventReceiverConfiguration.class})
-public class UacEventReceiverImplIT_Config {
+public class MessageIT_Config {
+
+  /** Spy on Service Activator Message End point */
+  @Bean
+  public CaseEventReceiverImpl caseEventReceiver() {
+    return Mockito.spy(new CaseEventReceiverImpl());
+  }
+
+  /** Spy on Service Activator Message End point */
+  @Bean
+  public CollectionExerciseEventReceiverImpl collectionExerciseEventReceiver() {
+    return Mockito.spy(new CollectionExerciseEventReceiverImpl());
+  }
+
+  /** Spy on Service Activator Message End point */
+  @Bean
+  public SurveyEventReceiverImpl surveyEventReceiver() {
+    return Mockito.spy(new SurveyEventReceiverImpl());
+  }
 
   /** Spy on Service Activator Message End point */
   @Bean
