--- src/test/java/uk/gov/ons/ctp/integration/rhsvc/message/impl/SurveyEventReceiverImplIT_Config.java
@@ -1,21 +1,0 @@
-package uk.gov.ons.ctp.integration.rhsvc.message.impl;
-
-import org.mockito.Mockito;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.Profile;
-import uk.gov.ons.ctp.integration.rhsvc.config.InboundEventIntegrationConfig;
-import uk.gov.ons.ctp.integration.rhsvc.event.impl.SurveyEventReceiverImpl;
-
-@Profile("mocked-connection-factory")
-@Configuration
-@Import({InboundEventIntegrationConfig.class, EventReceiverConfiguration.class})
-public class SurveyEventReceiverImplIT_Config {
-
-  /** Spy on Service Activator Message End point */
-  @Bean
-  public SurveyEventReceiverImpl surveyEventReceiver() {
-    return Mockito.spy(new SurveyEventReceiverImpl());
-  }
-}
